Switch to Connexion 3 framework
This is a huge PR being result of over a 100 commits
made by a number of people in ##36052 and #37638. It
switches to Connexion 3 as the driving backend
implementation for both - Airflow REST APIs and Flask
app that powers Airflow UI. It should be largely
backwards compatible when it comes to behaviour of
both APIs and Airflow Webserver views, however due to
decisions made by Connexion 3 maintainers, it changes
heavily the technology stack used under-the-hood:

1) Connexion App is an ASGI-compatible Open-API spec-first
   framework using ASGI as an interface between webserver
   and Python web application. ASGI is an asynchronous
   successor of WSGI.

2) Connexion itself is using Starlette to run asynchronous
   web services in Python.

3) We continue using gunicorn appliation server that still
   uses WSGI standard, which means that we can continue using
   Flask and we are usig standard Uvicorn ASGI webserver that
   converts the ASGI interface to WSGI interface of Gunicorn

Some of the problems handled in this PR

There were two problem was with session handling:

* the get_session_cookie - did not get the right cookie - it returned
  "session" string. The right fix was to change cookie_jar into
  cookie.jar because this is where apparently TestClient of starlette
  is holding the cookies (visible when you debug)

* The client does not accept "set_cookie" method - it accepts passing
  cookies via "cookies" dictionary - this is the usual httpx client
  - see  https://www.starlette.io/testclient/ - so we have to set
  cookie directly in the get method to try it out

Add "flask_client_with_login" for tests that neeed flask client

Some tests require functionality not available to Starlette test
client as they use Flask test client specific features - for those
we have an option to get flask test client instead of starlette
one.

Fix error handling for new connection 3 approach

Error handling for Connexion 3 integration needed to be reworked.

The way it behaves is much the same as it works in main:

* for API errors - we get application/problem+json responses
* for UI erros - we have rendered views
* for redirection - we have correct location header (it's been
  missing)
* the api error handled was not added as available middleware
  in the www tests

It should fix all test_views_base.py tests which were failing
on lack of location header for redirection.

Fix wrong response is tests_view_cluster_activity

The problem in the test was that Starlette Test Client opens a new
connection and start new session, while flask test client
uses the same database session. The test did not show data because
the data was not committed and session was not closed - which also
failed sqlite local tests with "database is locked" error.

Fix test_extra_links

The tests were failing again because the dagrun created was not
committed and session not closed. This worked with flask client
that used the same session accidentally but did not work with
test client from Starlette. Also it caused "database locked"
in sqlite / local tests.

Switch to non-deprecated auth manager

Fix to test_views_log.py

This PR partially fixes sessions and request parameter for
test_views_log. Some tests are still failing but for different reasons -
to be investigated.

Fix views_custom_user_views tests

The problem in those tests was that the check in security manager
was based on the assumption that the security manager was shared
between the client and test flask application - because they
were coming from the same flask app. But when we use starlette,
the call goes to a new process started and the user is deleted in
the database - so the shortcut of checking the security manager
did not work.

The change is that we are now checking if the user is deleted by
calling /users/show (we need a new users READ permission for that)
 - this way we go to the database and check if the user was indeed
deleted.

Fix test_task_instance_endpoint tests

There were two reasons for the test failed:

* when the Job was added to task instance, the task instance was not
  merged in session, which means that commit did not store the added
  Job

* some of the tests were expecting a call with specific session
  and they failed because session was different. Replacing the
  session with mock.ANY tells pytest that this parameter can be
  anything - we will have different session when when the call will
  be made with ASGI/Starlette

Fix parameter validation

* added default value for limit parameter across the board. Connexion
  3 does not like if the parameter had no default and we had not
    provided one - even if our custom decorated was adding it. Adding
  default value and updating our decorator to treat None as `default`
  fixed a number of problems where limits were not passed

* swapped openapi specification for /datasets/{uri} and /dataset/events.
  Since `{uri}` was defined first, connection matched `events` with
  `{uri}` and chose parameter definitions from `{uri}` not events

Fix test_log_enpoint tests

The problem here was that some sessions should be committed/closed but
also in order to run it standalone we wanted to create log templates
in the database - as it relied implcitly on log templates created by
other tests.

Fix test_views_dagrun, test_views_tasks and test_views_log

Fixed by switching to use flask client for testing rather than
starlette. Starlette client in this case has some side effects that are
also impacting Sqlite's session being created in a different
thread and deleted with close_all_sessions fixture.

Fix test_views_dagrun

Fixed by switching to use flask client for testing rather than
starlette. Starlette client in this case has some side effects that are
also impacting Sqlite's session being created in a different
thread and deleted with close_all_sessions fixture.


Better API initialization including vending of API specification.

The way paths are added and initialized is better (for example
FAB contributes their path via new method in Auth Manager.

This also add back-compatibility to FAB auth manaager to continue
working on Airflow 2.9.

+++ tests/providers/fab/auth_manager/api_endpoints/test_role_and_permission_endpoint.py
@@ -28,7 +28,6 @@ with ignore_provider_compatibility_error("2.9.0+", __file__):
 
 from airflow.security import permissions
 from tests.test_utils.api_connexion_utils import (
-    assert_401,
     create_role,
     create_user,
     delete_role,
@@ -40,9 +39,9 @@ pytestmark = pytest.mark.db_test
 
 @pytest.fixture(scope="module")
 def configured_app(minimal_app_for_auth_api):
-    app = minimal_app_for_auth_api
+    connexion_app = minimal_app_for_auth_api
     create_user(
-        app,  # type: ignore
+        connexion_app.app,  # type: ignore
         username="test",
         role_name="Test",
         permissions=[
@@ -53,58 +52,55 @@ def configured_app(minimal_app_for_auth_api):
             (permissions.ACTION_CAN_READ, permissions.RESOURCE_ACTION),
         ],
     )
-    create_user(app, username="test_no_permissions", role_name="TestNoPermissions")  # type: ignore
-    yield app
+    create_user(connexion_app.app, username="test_no_permissions", role_name="TestNoPermissions")  # type: ignore
+    yield connexion_app
 
-    delete_user(app, username="test")  # type: ignore
-    delete_user(app, username="test_no_permissions")  # type: ignore
+    delete_user(connexion_app.app, username="test")  # type: ignore
+    delete_user(connexion_app.app, username="test_no_permissions")  # type: ignore
 
 
 class TestRoleEndpoint:
     @pytest.fixture(autouse=True)
     def setup_attrs(self, configured_app) -> None:
-        self.app = configured_app
-        self.client = self.app.test_client()  # type:ignore
+        self.connexion_app = configured_app
+        self.flask_app = self.connexion_app.app
+        self.client = self.connexion_app.test_client()  # type:ignore
 
     def teardown_method(self):
         """
         Delete all roles except these ones.
         Test and TestNoPermissions are deleted by delete_user above
         """
-        session = self.app.appbuilder.get_session
+        session = self.flask_app.appbuilder.get_session
         existing_roles = set(EXISTING_ROLES)
         existing_roles.update(["Test", "TestNoPermissions"])
         roles = session.query(Role).filter(~Role.name.in_(existing_roles)).all()
         for role in roles:
-            delete_role(self.app, role.name)
+            delete_role(self.flask_app, role.name)
 
 
 class TestGetRoleEndpoint(TestRoleEndpoint):
     def test_should_response_200(self):
-        response = self.client.get("/auth/fab/v1/roles/Admin", environ_overrides={"REMOTE_USER": "test"})
+        response = self.client.get("/auth/fab/v1/roles/Admin", headers={"REMOTE_USER": "test"})
         assert response.status_code == 200
-        assert response.json["name"] == "Admin"
+        assert response.json()["name"] == "Admin"
 
     def test_should_respond_404(self):
-        response = self.client.get(
-            "/auth/fab/v1/roles/invalid-role", environ_overrides={"REMOTE_USER": "test"}
-        )
+        response = self.client.get("/auth/fab/v1/roles/invalid-role", headers={"REMOTE_USER": "test"})
         assert response.status_code == 404
         assert {
             "detail": "Role with name 'invalid-role' was not found",
             "status": 404,
             "title": "Role not found",
             "type": EXCEPTIONS_LINK_MAP[404],
-        } == response.json
+        } == response.json()
 
     def test_should_raises_401_unauthenticated(self):
         response = self.client.get("/auth/fab/v1/roles/Admin")
-        assert_401(response)
+        assert response.status_code == 401
 
     def test_should_raise_403_forbidden(self):
-        response = self.client.get(
-            "/auth/fab/v1/roles/Admin", environ_overrides={"REMOTE_USER": "test_no_permissions"}
-        )
+        response = self.client.get("/auth/fab/v1/roles/Admin", headers={"REMOTE_USER": "test_no_permissions"})
         assert response.status_code == 403
 
     @pytest.mark.parametrize(
@@ -119,30 +115,26 @@ class TestGetRoleEndpoint(TestRoleEndpoint):
 
 class TestGetRolesEndpoint(TestRoleEndpoint):
     def test_should_response_200(self):
-        response = self.client.get("/auth/fab/v1/roles", environ_overrides={"REMOTE_USER": "test"})
+        response = self.client.get("/auth/fab/v1/roles", headers={"REMOTE_USER": "test"})
         assert response.status_code == 200
         existing_roles = set(EXISTING_ROLES)
         existing_roles.update(["Test", "TestNoPermissions"])
-        assert response.json["total_entries"] == len(existing_roles)
-        roles = {role["name"] for role in response.json["roles"]}
+        assert response.json()["total_entries"] == len(existing_roles)
+        roles = {role["name"] for role in response.json()["roles"]}
         assert roles == existing_roles
 
     def test_should_raises_401_unauthenticated(self):
         response = self.client.get("/auth/fab/v1/roles")
-        assert_401(response)
+        assert response.status_code == 401
 
     def test_should_raises_400_for_invalid_order_by(self):
-        response = self.client.get(
-            "/auth/fab/v1/roles?order_by=invalid", environ_overrides={"REMOTE_USER": "test"}
-        )
+        response = self.client.get("/auth/fab/v1/roles?order_by=invalid", headers={"REMOTE_USER": "test"})
         assert response.status_code == 400
         msg = "Ordering with 'invalid' is disallowed or the attribute does not exist on the model"
-        assert response.json["detail"] == msg
+        assert response.json()["detail"] == msg
 
     def test_should_raise_403_forbidden(self):
-        response = self.client.get(
-            "/auth/fab/v1/roles", environ_overrides={"REMOTE_USER": "test_no_permissions"}
-        )
+        response = self.client.get("/auth/fab/v1/roles", headers={"REMOTE_USER": "test_no_permissions"})
         assert response.status_code == 403
 
     @pytest.mark.parametrize(
@@ -178,33 +170,31 @@ class TestGetRolesEndpointPaginationandFilter(TestRoleEndpoint):
         ],
     )
     def test_can_handle_limit_and_offset(self, url, expected_roles):
-        response = self.client.get(url, environ_overrides={"REMOTE_USER": "test"})
+        response = self.client.get(url, headers={"REMOTE_USER": "test"})
         assert response.status_code == 200
         existing_roles = set(EXISTING_ROLES)
         existing_roles.update(["Test", "TestNoPermissions"])
-        assert response.json["total_entries"] == len(existing_roles)
-        roles = [role["name"] for role in response.json["roles"] if role]
+        assert response.json()["total_entries"] == len(existing_roles)
+        roles = [role["name"] for role in response.json()["roles"] if role]
 
         assert roles == expected_roles
 
 
 class TestGetPermissionsEndpoint(TestRoleEndpoint):
     def test_should_response_200(self):
-        response = self.client.get("/auth/fab/v1/permissions", environ_overrides={"REMOTE_USER": "test"})
-        actions = {i[0] for i in self.app.appbuilder.sm.get_all_permissions() if i}
+        response = self.client.get("/auth/fab/v1/permissions", headers={"REMOTE_USER": "test"})
+        actions = {i[0] for i in self.flask_app.appbuilder.sm.get_all_permissions() if i}
         assert response.status_code == 200
-        assert response.json["total_entries"] == len(actions)
-        returned_actions = {perm["name"] for perm in response.json["actions"]}
+        assert response.json()["total_entries"] == len(actions)
+        returned_actions = {perm["name"] for perm in response.json()["actions"]}
         assert actions == returned_actions
 
     def test_should_raises_401_unauthenticated(self):
         response = self.client.get("/auth/fab/v1/permissions")
-        assert_401(response)
+        assert response.status_code == 401
 
     def test_should_raise_403_forbidden(self):
-        response = self.client.get(
-            "/auth/fab/v1/permissions", environ_overrides={"REMOTE_USER": "test_no_permissions"}
-        )
+        response = self.client.get("/auth/fab/v1/permissions", headers={"REMOTE_USER": "test_no_permissions"})
         assert response.status_code == 403
 
     @pytest.mark.parametrize(
@@ -223,11 +213,9 @@ class TestPostRole(TestRoleEndpoint):
             "name": "Test2",
             "actions": [{"resource": {"name": "Connections"}, "action": {"name": "can_create"}}],
         }
-        response = self.client.post(
-            "/auth/fab/v1/roles", json=payload, environ_overrides={"REMOTE_USER": "test"}
-        )
+        response = self.client.post("/auth/fab/v1/roles", json=payload, headers={"REMOTE_USER": "test"})
         assert response.status_code == 200
-        role = self.app.appbuilder.sm.find_role("Test2")
+        role = self.flask_app.appbuilder.sm.find_role("Test2")
         assert role is not None
 
     @pytest.mark.parametrize(
@@ -296,11 +284,9 @@ class TestPostRole(TestRoleEndpoint):
         ],
     )
     def test_post_should_respond_400_for_invalid_payload(self, payload, error_message):
-        response = self.client.post(
-            "/auth/fab/v1/roles", json=payload, environ_overrides={"REMOTE_USER": "test"}
-        )
+        response = self.client.post("/auth/fab/v1/roles", json=payload, headers={"REMOTE_USER": "test"})
         assert response.status_code == 400
-        assert response.json == {
+        assert response.json() == {
             "detail": error_message,
             "status": 400,
             "title": "Bad Request",
@@ -312,11 +298,9 @@ class TestPostRole(TestRoleEndpoint):
             "name": "Test",
             "actions": [{"resource": {"name": "Connections"}, "action": {"name": "can_create"}}],
         }
-        response = self.client.post(
-            "/auth/fab/v1/roles", json=payload, environ_overrides={"REMOTE_USER": "test"}
-        )
+        response = self.client.post("/auth/fab/v1/roles", json=payload, headers={"REMOTE_USER": "test"})
         assert response.status_code == 409
-        assert response.json == {
+        assert response.json() == {
             "detail": "Role with name 'Test' already exists; please update with the PATCH endpoint",
             "status": 409,
             "title": "Conflict",
@@ -332,7 +316,7 @@ class TestPostRole(TestRoleEndpoint):
             },
         )
 
-        assert_401(response)
+        assert response.status_code == 401
 
     def test_should_raise_403_forbidden(self):
         response = self.client.post(
@@ -341,7 +325,7 @@ class TestPostRole(TestRoleEndpoint):
                 "name": "mytest2",
                 "actions": [{"resource": {"name": "Connections"}, "action": {"name": "can_create"}}],
             },
-            environ_overrides={"REMOTE_USER": "test_no_permissions"},
+            headers={"REMOTE_USER": "test_no_permissions"},
         )
         assert response.status_code == 403
 
@@ -361,20 +345,16 @@ class TestPostRole(TestRoleEndpoint):
 
 class TestDeleteRole(TestRoleEndpoint):
     def test_delete_should_respond_204(self, session):
-        role = create_role(self.app, "mytestrole")
-        response = self.client.delete(
-            f"/auth/fab/v1/roles/{role.name}", environ_overrides={"REMOTE_USER": "test"}
-        )
+        role = create_role(self.flask_app, "mytestrole")
+        response = self.client.delete(f"/auth/fab/v1/roles/{role.name}", headers={"REMOTE_USER": "test"})
         assert response.status_code == 204
         role_obj = session.query(Role).filter(Role.name == role.name).all()
         assert len(role_obj) == 0
 
     def test_delete_should_respond_404(self):
-        response = self.client.delete(
-            "/auth/fab/v1/roles/invalidrolename", environ_overrides={"REMOTE_USER": "test"}
-        )
+        response = self.client.delete("/auth/fab/v1/roles/invalidrolename", headers={"REMOTE_USER": "test"})
         assert response.status_code == 404
-        assert response.json == {
+        assert response.json() == {
             "detail": "Role with name 'invalidrolename' was not found",
             "status": 404,
             "title": "Role not found",
@@ -384,11 +364,11 @@ class TestDeleteRole(TestRoleEndpoint):
     def test_should_raises_401_unauthenticated(self):
         response = self.client.delete("/auth/fab/v1/roles/test")
 
-        assert_401(response)
+        assert response.status_code == 401
 
     def test_should_raise_403_forbidden(self):
         response = self.client.delete(
-            "/auth/fab/v1/roles/test", environ_overrides={"REMOTE_USER": "test_no_permissions"}
+            "/auth/fab/v1/roles/test", headers={"REMOTE_USER": "test_no_permissions"}
         )
         assert response.status_code == 403
 
@@ -398,7 +378,7 @@ class TestDeleteRole(TestRoleEndpoint):
         indirect=["set_auto_role_public"],
     )
     def test_with_auth_role_public_set(self, set_auto_role_public, expected_status_code):
-        role = create_role(self.app, "mytestrole")
+        role = create_role(self.flask_app, "mytestrole")
         response = self.client.delete(f"/auth/fab/v1/roles/{role.name}")
         assert response.status_code == expected_status_code
 
@@ -419,17 +399,17 @@ class TestPatchRole(TestRoleEndpoint):
         ],
     )
     def test_patch_should_respond_200(self, payload, expected_name, expected_actions):
-        role = create_role(self.app, "mytestrole")
+        role = create_role(self.flask_app, "mytestrole")
         response = self.client.patch(
-            f"/auth/fab/v1/roles/{role.name}", json=payload, environ_overrides={"REMOTE_USER": "test"}
+            f"/auth/fab/v1/roles/{role.name}", json=payload, headers={"REMOTE_USER": "test"}
         )
         assert response.status_code == 200
-        assert response.json["name"] == expected_name
-        assert response.json["actions"] == expected_actions
+        assert response.json()["name"] == expected_name
+        assert response.json()["actions"] == expected_actions
 
     def test_patch_should_update_correct_roles_permissions(self):
-        create_role(self.app, "role_to_change")
-        create_role(self.app, "already_exists")
+        create_role(self.flask_app, "role_to_change")
+        create_role(self.flask_app, "already_exists")
 
         response = self.client.patch(
             "/auth/fab/v1/roles/role_to_change",
@@ -437,16 +417,16 @@ class TestPatchRole(TestRoleEndpoint):
                 "name": "already_exists",
                 "actions": [{"action": {"name": "can_delete"}, "resource": {"name": "XComs"}}],
             },
-            environ_overrides={"REMOTE_USER": "test"},
+            headers={"REMOTE_USER": "test"},
         )
         assert response.status_code == 200
 
-        updated_permissions = self.app.appbuilder.sm.find_role("role_to_change").permissions
+        updated_permissions = self.flask_app.appbuilder.sm.find_role("role_to_change").permissions
         assert len(updated_permissions) == 1
         assert updated_permissions[0].resource.name == "XComs"
         assert updated_permissions[0].action.name == "can_delete"
 
-        assert len(self.app.appbuilder.sm.find_role("already_exists").permissions) == 0
+        assert len(self.flask_app.appbuilder.sm.find_role("already_exists").permissions) == 0
 
     @pytest.mark.parametrize(
         "update_mask, payload, expected_name, expected_actions",
@@ -474,27 +454,27 @@ class TestPatchRole(TestRoleEndpoint):
     def test_patch_should_respond_200_with_update_mask(
         self, update_mask, payload, expected_name, expected_actions
     ):
-        role = create_role(self.app, "mytestrole")
+        role = create_role(self.flask_app, "mytestrole")
         assert role.permissions == []
         response = self.client.patch(
             f"/auth/fab/v1/roles/{role.name}{update_mask}",
             json=payload,
-            environ_overrides={"REMOTE_USER": "test"},
+            headers={"REMOTE_USER": "test"},
         )
         assert response.status_code == 200
-        assert response.json["name"] == expected_name
-        assert response.json["actions"] == expected_actions
+        assert response.json()["name"] == expected_name
+        assert response.json()["actions"] == expected_actions
 
     def test_patch_should_respond_400_for_invalid_fields_in_update_mask(self):
-        role = create_role(self.app, "mytestrole")
+        role = create_role(self.flask_app, "mytestrole")
         payload = {"name": "testme"}
         response = self.client.patch(
             f"/auth/fab/v1/roles/{role.name}?update_mask=invalid_name",
             json=payload,
-            environ_overrides={"REMOTE_USER": "test"},
+            headers={"REMOTE_USER": "test"},
         )
         assert response.status_code == 400
-        assert response.json["detail"] == "'invalid_name' in update_mask is unknown"
+        assert response.json()["detail"] == "'invalid_name' in update_mask is unknown"
 
     @pytest.mark.parametrize(
         "payload, expected_error",
@@ -547,14 +527,14 @@ class TestPatchRole(TestRoleEndpoint):
         ],
     )
     def test_patch_should_respond_400_for_invalid_update(self, payload, expected_error):
-        role = create_role(self.app, "mytestrole")
+        role = create_role(self.flask_app, "mytestrole")
         response = self.client.patch(
             f"/auth/fab/v1/roles/{role.name}",
             json=payload,
-            environ_overrides={"REMOTE_USER": "test"},
+            headers={"REMOTE_USER": "test"},
         )
         assert response.status_code == 400
-        assert response.json["detail"] == expected_error
+        assert response.json()["detail"] == expected_error
 
     def test_should_raises_401_unauthenticated(self):
         response = self.client.patch(
@@ -565,7 +545,7 @@ class TestPatchRole(TestRoleEndpoint):
             },
         )
 
-        assert_401(response)
+        assert response.status_code == 401
 
     def test_should_raise_403_forbidden(self):
         response = self.client.patch(
@@ -574,7 +554,7 @@ class TestPatchRole(TestRoleEndpoint):
                 "name": "mytest2",
                 "actions": [{"resource": {"name": "Connections"}, "action": {"name": "can_create"}}],
             },
-            environ_overrides={"REMOTE_USER": "test_no_permissions"},
+            headers={"REMOTE_USER": "test_no_permissions"},
         )
         assert response.status_code == 403
 
@@ -584,7 +564,7 @@ class TestPatchRole(TestRoleEndpoint):
         indirect=["set_auto_role_public"],
     )
     def test_with_auth_role_public_set(self, set_auto_role_public, expected_status_code):
-        role = create_role(self.app, "mytestrole")
+        role = create_role(self.flask_app, "mytestrole")
         response = self.client.patch(
             f"/auth/fab/v1/roles/{role.name}",
             json={"name": "mytest"},
